third_party/WebKit/LayoutTests/editing/execCommand/crash-indenting-list-item.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contentEditable><ul><li>|</li></ul></div>',
        'indent',
        '<div contenteditable><ul><ul><li></li></ul>|</ul></div>'),
        'This is crash test');
</script>
